Ora2ES: more accurate handling of case when no new data found.
Sending empty file to ES produces a null_pointer_exception error.

+++ Utils/Elasticsearch/process.sh
@@ -27,7 +27,11 @@ convertDataToESFormat() {
 }
 
 getDataFromOracle > $ORACLE_OUT 
-convertDataToESFormat < $ORACLE_OUT >$ES_IN 
-$(dirname $0)/tools/load_data.sh $ES_IN #step 3 - load data to ES
-echo OFFSET="'${NEW_OFFSET}'" > "$OFFSET_FILE"
+if [ -s "$ORACLE_OUT" ]; then
+  convertDataToESFormat < $ORACLE_OUT >$ES_IN 
+  $(dirname $0)/tools/load_data.sh $ES_IN #step 3 - load data to ES
+  echo OFFSET="'${NEW_OFFSET}'" > "$OFFSET_FILE"
+else
+  log "No data changed since last offset."
+fi
 rm -f $ORACLE_OUT $ES_IN
